requirements.txt
Pillow
ultralytics
tqdm
datasets